+++ templates/login.html
@@ -1,7 +1,8 @@
-<!-- Admin initialization page -->
-{% extends "base.html" %} {% block head %}
+{% extends "base.html" %} {% block head %} {% if backup %}
+<title>Backup Login</title>
+{% else %}
 <title>Login</title>
-{% include 'scripts/toast.html' %} {% endblock %} {% block body %}
+{% endif %} {% include 'scripts/toast.html' %} {% endblock %} {% block body %}
 <div class="h-screen w-full flex items-center justify-center">
   {% block error_toast %}
   <div id="message" class="hidden">
@@ -20,7 +21,16 @@
     hx-target="#message"
   >
     <input type="hidden" name="redirect_uri" value="{{ redirect_uri }}" />
+    {% if backup %}
+    <h1 class="text-3xl">Backup Login</h1>
+    <p class="text-error">
+      This instance has OIDC enabled. This login can only be used by the root
+      admin.
+    </p>
+    {% else %}
     <h1 class="text-3xl">Login</h1>
+    {% endif %}
+
     <label for="username">Username</label>
     <input
       id="username"
