+++ templates/settings_page/notifications.html
@@ -3,7 +3,7 @@
 {% endblock %} {% block content %}
 
 <div class="flex flex-col">
-  <h2 class="text-lg">Notications</h2>
+  <h2 class="text-lg">Notifications</h2>
 
   <form
     id="add-notification-form"
